refactor(ATT-425): split resourceDetailsScreen god file into focused modules
Re-split the resourceDetailsScreen.cpp god file on top of current main
(which added the maintenance panel and reworked forms into a multi-page
next/back flow). Partition by concern, methods stay members of
ResourceDetailsScreen (header unchanged):
- resourceDetailsScreen.cpp: core lifecycle + access/maintenance state
- resourceDetailsForms.cpp: dynamic multi-page forms modal + select fields
- resourceDetailsProjects.cpp: projects modal + pagination
- resourceDetailsActionFeedback.cpp: action overlay + success toasts
- resourceDetailsSession.cpp: elapsed-time / session timeout display

Method bodies copied verbatim from main (multiset-verified, zero lost
lines); file-local statics moved to their owning translation unit.
Bump firmware version 1.3.13 -> 1.3.14 for OTA.
